Split cluster into its own package
